+++ src/routes/(main)/games/create/+page.server.ts
@@ -13,8 +13,8 @@ const schema = z.object({
 		.regex(
 			/^(http:\/\/www\.|https:\/\/www\.|http:\/\/|https:\/\/)?[a-z0-9]+([\-\.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?|^((http:\/\/www\.|https:\/\/www\.|http:\/\/|https:\/\/)?([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])$/
 		),
-	serverPort: z.number().min(25565).max(65535).default(53640),
-	maxPlayers: z.number().min(1).max(99).default(10),
+	serverPort: z.number().int().min(25565).max(65535).default(53640),
+	maxPlayers: z.number().int().min(1).max(99).default(10),
 	privateServer: z.boolean().optional(),
 })
